BL-6271 new books never have copyright or versionAcknowledgements
